+++ src/variables/aboutConsts.js
@@ -1,8 +1,8 @@
 const GLOBAL_NAME = "BubbleOS";
 const AUTHOR = "Arnav Thorat";
 
-const VERSION = "0.4.8";
-const BUILD = "48";
+const VERSION = "0.5.4";
+const BUILD = "54";
 const IN_BETA = true;
 
 module.exports = { GLOBAL_NAME, AUTHOR, VERSION, BUILD, IN_BETA };